First implementation of iSite algorithm
The iSite algorithm is now implemented. More testing should be done to confirm
robustness. Still to be implemented is deletion of a node with no iSites/edges.

Also implemented an initial draft of a check for self loops in duplication.
Much more work still needs to be done on this.